Change the interface to use channels
Change the interface layer to use Tokio's channels instead of a queue,
as the blocking Mutex causes task execution to stop.

+++ backend/messages/src/client.rs
@@ -17,7 +17,7 @@ pub enum ClientMessage {
     /// Message to send Job Config
     JobConfig {
         /// Job timeout
-        timeout: u8,
+        timeout: i32,
         /// Types of each column in dataset for job
         column_types: Vec<String>,
     },
